nodesign: require interface in the API
a common source of bugs (happened in the past) is the
mismatch between the namespace/name (name/namespace) string
arguments, leading to hard to debug issues.

Since both NRT objects and pod objects are expected to
implement GetNamespace() and GetName(), use a interface
to clarify and enforce the API constraint, avoiding
any possible source of confusion.

Signed-off-by: Francesco Romani <fromani@redhat.com>